MarketDataFetcher/requirements.txt
fastapi
uvicorn[standard]
requests
kafka-python-ng
pydantic
typing


